added automation shell script for video rendering | implemented 3d anaglyph and non-anaglyph menu options
