+++ python/bertini_real/__init__.py
@@ -8,7 +8,6 @@ import bertini_real.glumpyplotter
 import bertini_real.tmesh
 import bertini_real.anaglypy
 
-
 def gather_and_plot():
     bertini_real.data.gather()
     return bertini_real.plot.plot()
