+++ python/bertini_real/anaglypy/__init__.py
@@ -1,23 +1,4 @@
-import os
-import bertini_real
-
-
-def rotate_xyz():
-    # script_path = os.path.dirname(
-    #     bertini_real.__file__) + "/anaglypy/rotate_xyz.py"
-    # os.system("blender -noaudio -b -P " + script_path)
-    print("rotate xyz")
-
-
-def rotate_z():
-    # script_path = os.path.dirname(
-    #     bertini_real.__file__) + "/anaglypy/rotate_z.py"
-    # os.system("blender -noaudio -b -P " + script_path)
-    print("rotate z")
 
 
 if __name__ == "__main__":
-    rotate_xyz()
-    rotate_z()
-
-
+    import anaglypy
