+++ python/bertini_real/anaglypy/anaglypy.sh
@@ -5,11 +5,9 @@ for D in *; do
 
     	if grep -Fxq "${D}" surfaces.txt
     	then
-    		cd "${D}" # go to surface directory
+    		cd "${D}" # navigate to surface directory
     		blender -b -P ../anaglypy.py < ../options.txt
     		cd ..
-    	# else
-    		# do nothing
     	fi
     fi
 done
